AIExamSystem-main/python-backend/requirements.txt
google-generativeai==0.3.1
